Add files via upload
The most important aspect of any Data Science approach is how the information is processed. When we talk about developing insights out of data it is basically digging out the possibilities. Those possibilities in Data Science are known as Statistical Analysis.

Most of us wonder how can data in the form of text, images, videos, and other highly unstructured formats get easily processed by Machine Learning models. But, the truth is we actually convert that data into a numerical form which is not exactly our data but the numerical equivalent of it. So, this brings us to the very important aspect of Data Science.

With data in numerical format, it provides us with infinite possibilities to understand the information out it. Statistics acts as a pathway to understand your data and process that for successful results. Not only the power of statistics is limited to understanding the data it also provides methods to measure the success of our insights, getting different approaches for the same problem, getting the right mathematical approach for your data.